requirements.txt
tweepy==*
python-dotenv==*